data_gen_qwenvl/requirements.txt
numpy
torch>=2.0.0
transformers
scikit-learn
underthesea
pyvi
tqdm
PyYAML
jinja2
tokenizers>=0.11.1,!=0.11.3
torchvision
bitsandbytes 
accelerate
einops
timm